+++ src/main/java/com/mindata/blockchain/core/requestbody/InstructionBody.java
@@ -7,7 +7,7 @@ public class InstructionBody {
     /**
      * 指令的操作，增删改
      */
-    private int operation;
+    private byte operation;
     /**
      * 操作的model对象，对应一张表
      */
@@ -44,11 +44,11 @@ public class InstructionBody {
         this.publicKey = publicKey;
     }
 
-    public int getOperation() {
+    public byte getOperation() {
         return operation;
     }
 
-    public void setOperation(int operation) {
+    public void setOperation(byte operation) {
         this.operation = operation;
     }
